dziala 1

+++ app/src/main/java/pl/edu/pwr/kotlin/mytasks/adapters/TasksAdapter.kt
@@ -33,7 +33,7 @@ class TasksAdapter(private val mContext: Context, private val mTasks: ArrayList<
     override fun onBindViewHolder(holder: MyViewHolder, position: Int) {
         val name = mTasks[position].name
         holder.nameStr.text=name;
-        val color = when(name.get(0) % 5){
+        val color = when(name.get(0).toInt() % 5){
             0 -> color(R.color.task0)
             1 -> color(R.color.task1)
             2 -> color(R.color.task2)
